Fix setuptools arg for install
requires isn't a kwarg for setuptools.setup, install_requires is.

There could be some further cleanup to the installation requirements to
be more consistent. setup.py is compiling its own list instead of using
requirements.txt. I'd probably remove requirements.txt altogether, but
this is a direct resolution to the import failure that occurs if the
requirements aren't installed separately from the package itself.